+++ src/frontend/packages/field-components/src/index.js
@@ -1,3 +1,5 @@
 export { default as AvatarWithLabelField } from './AvatarWithLabelField';
+export { default as ReferenceArrayField } from './ReferenceArrayField';
+export { default as ReferenceField } from './ReferenceField';
 export { default as QuickAppendReferenceArrayField } from './QuickAppendReferenceArrayField/QuickAppendReferenceArrayField';
 export { default as MultiUrlField } from './MultiUrlField';
