+++ src/frontend/packages/input-components/src/index.js
@@ -1,3 +1,5 @@
 export { default as MultiServerAutocompleteArrayInput } from './MultiServerAutocompleteArrayInput';
 export { default as MultiServerAutocompleteInput } from './MultiServerAutocompleteInput';
 export { default as MultiLinesInput } from './MultiLinesInput';
+export { default as ReferenceArrayInput } from './ReferenceArrayInput';
+export { default as ReferenceInput } from './ReferenceInput';
