+++ src/frontend/packages/semantic-data-provider/src/index.js
@@ -1,13 +1,8 @@
-export { default as ImageField } from './fields/ImageField';
-export { default as UriArrayField, default as ReferenceArrayField } from './fields/ReferenceArrayField';
-export { default as ReferenceField } from './fields/ReferenceField';
-export { default as FilterHandler } from './dataHandling/FilterHandler';
-export { default as GroupedReferenceHandler } from './dataHandling/GroupedReferenceHandler';
+export { default as dataProvider } from './dataProvider/dataProvider';
+export { default as httpClient } from './httpClient';
 
-export { default as DateTimeInput } from './inputs/DateTimeInput';
-export { default as UriArrayInput, default as ReferenceArrayInput } from './inputs/ReferenceArrayInput';
-export { default as ReferenceInput } from './inputs/ReferenceInput';
-export { default as ReificationArrayInput } from './inputs/ReificationArrayInput';
+export { default as buildSparqlQuery } from './dataProvider/utils/buildSparqlQuery';
+export { default as buildBlankNodesQuery } from './dataProvider/utils/buildBlankNodesQuery';
 
 export { default as useGetExternalLink } from './hooks/useGetExternalLink';
 export { default as useContainers } from './hooks/useContainers';
@@ -16,8 +11,6 @@ export { default as useDataModel } from './hooks/useDataModel';
 export { default as useDataModels } from './hooks/useDataModels';
 export { default as useDataServers } from './hooks/useDataServers';
 
-export { default as dataProvider } from './dataProvider/dataProvider';
-export { default as httpClient } from './httpClient';
-
-export { default as buildSparqlQuery } from './dataProvider/utils/buildSparqlQuery';
-export { default as buildBlankNodesQuery } from './dataProvider/utils/buildBlankNodesQuery';
+export { default as FilterHandler } from './reification/FilterHandler';
+export { default as GroupedReferenceHandler } from './reification/GroupedReferenceHandler';
+export { default as ReificationArrayInput } from './reification/ReificationArrayInput';
