+++ src/frontend/packages/semantic-data-provider/src/reification/GroupedReferenceHandler.js
